fix: mydocument and uploaddocument bugs (cant upload)
